\name{heyhey-package}
\alias{heyhey-package}
\alias{heyhey}
\docType{package}
\title{
\packageTitle{heyhey}
}
\description{
\packageDescription{heyhey}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{heyhey}
\packageIndices{heyhey}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{heyhey}

Maintainer: \packageMaintainer{heyhey}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
